+++ lib/hexagon.dart
@@ -8,6 +8,7 @@ library;
 export 'src/grid/coordinates.dart';
 export 'src/grid/hexagon_grid.dart';
 export 'src/grid/hexagon_offset_grid.dart';
+export 'src/hexagon_border.dart';
 export 'src/hexagon_clipper.dart';
 export 'src/hexagon_painter.dart';
 export 'src/hexagon_path_builder.dart';